+++ ta4j-core/src/main/java/org/ta4j/core/indicators/bollinger/BollingerBandsMiddleIndicator.java
@@ -22,30 +22,30 @@
  */
 package org.ta4j.core.indicators.bollinger;
 
-import org.ta4j.core.Decimal;
+
 import org.ta4j.core.Indicator;
 import org.ta4j.core.indicators.CachedIndicator;
 
 /**
  * Buy - Occurs when the price line cross from down to up de Bollinger Band Low.
  * Sell - Occurs when the price line cross from up to down de Bollinger Band High.
- * 
+ *
  */
-public class BollingerBandsMiddleIndicator extends CachedIndicator<Decimal> {
+public class BollingerBandsMiddleIndicator extends CachedIndicator<Double> {
 
-    private final Indicator<Decimal> indicator;
+    private final Indicator<Double> indicator;
 
-    public BollingerBandsMiddleIndicator(Indicator<Decimal> indicator) {
+    public BollingerBandsMiddleIndicator(Indicator<Double> indicator) {
         super(indicator);
         this.indicator = indicator;
     }
 
     @Override
-    protected Decimal calculate(int index) {
+    protected Double calculate(int index) {
         return indicator.getValue(index);
     }
 
-    public Indicator<Decimal> getIndicator() {
+    public Indicator<Double> getIndicator() {
         return indicator;
     }
